[worker] extracted chain_relay methods from integration tests and incorporated chain_relay in worker. `demo_private_tx` run successfully.
