+++ client/Cargo.toml
@@ -21,6 +21,7 @@ blake2-rfc      = { version = "0.2.18", default-features = false}
 
 [dependencies.substrate-api-client]
 git = "https://github.com/scs/substrate-api-client"
+branch = "cl-subscribe-finalized-heads"
 
 [dependencies.serde]
 features = ["derive"]
