+++ enclave/Cargo.toml
@@ -101,6 +101,7 @@ default-features = false
 
 [dependencies.substrate-api-client]
 git = "https://github.com/scs/substrate-api-client"
+branch = "cl-subscribe-finalized-heads"
 default-features = false
 features = ["full_crypto"]
