+++ substratee-node-calls/Cargo.toml
@@ -18,6 +18,7 @@ package = "substratee-node-runtime"
 
 [dependencies.substrate-api-client]
 git = "https://github.com/scs/substrate-api-client"
+branch = "cl-subscribe-finalized-heads"
 
 [dependencies.sp-core]
 version = '2.0.0-alpha.7'
