+++ worker/Cargo.toml
@@ -36,6 +36,7 @@ sgx_crypto_helper 		= { rev = "v1.1.2", git = "https://github.com/apache/teaclav
 
 [dependencies.substrate-api-client]
 git = "https://github.com/scs/substrate-api-client"
+branch = "cl-subscribe-finalized-heads"
 
 [dependencies.substratee-node-calls]
 path = "../substratee-node-calls"
